CLI-839 Move terminal rendering into TerminalConsole
