+++ src/core/ui/types.ts
@@ -41,6 +41,15 @@ export interface PhaseItem {
   subItems?: string[];
 }
 
+export function phaseItem(
+  text: string,
+  status: StepStatus,
+  detail?: string,
+  subItems?: string[],
+): PhaseItem {
+  return { text, status, detail, subItems };
+}
+
 export interface NoteOptions {
   borderColor?: ColorFn;
   titleColor?: ColorFn;
@@ -54,4 +63,37 @@ export interface PhaseOptions {
 
 export interface LogOptions {
   color?: ColorFn;
+}
+
+export interface SelectOption<T> {
+  value: T;
+  label: string;
+}
+
+export interface MultiSelectOption<T> {
+  value: T;
+  label: string;
+  /** Shown but not toggleable (e.g. a repo that's already imported) — rendered like an at-cap row. */
+  disabled?: boolean;
+}
+
+export interface MultiSelectPromptOptions<T> {
+  /** True while more options can be revealed (e.g. a paginated collection's `hasMore`). */
+  hasMore?: () => boolean;
+  /**
+   * Reveal more options, returning (or resolving to) the full updated list (existing options
+   * plus newly revealed ones). May return a `Promise` when revealing more requires a network
+   * call (e.g. `RepositoryCollection.loadMore`) — the "Load more" row shows a loading indicator
+   * while it's in flight.
+   */
+  onLoadMore?: () => MultiSelectOption<T>[] | Promise<MultiSelectOption<T>[]>;
+  /** Max number of selections. Defaults to 20; pass `Infinity` for no cap. */
+  maxSelected?: number;
+  /**
+   * True total option count, for the "N/total selected" counter. Defaults to the currently
+   * revealed `options.length`, which undercounts while paginated options remain unrevealed —
+   * pass this when the caller's backing collection knows the real total up front (e.g.
+   * `RepositoryCollection.length`).
+   */
+  total?: () => number;
 }
