+++ tests/unit/core/ui/concurrent-progress.test.ts
@@ -22,8 +22,10 @@ import { describe, expect, it, spyOn } from 'bun:test';
 
 import { ImportProgress } from '@/commands/import/import-progress.ts';
 import { ConcurrentProgress } from '@/core/ui/components/concurrent-progress.ts';
+import { TerminalConsole } from '@/core/ui/terminal-console.ts';
 
 const ITEMS = ['alpha', 'beta', 'gamma'];
+const term = new TerminalConsole();
 
 function captureStdout(fn: () => void): string {
   const chunks: string[] = [];
@@ -40,14 +42,14 @@ function captureStdout(fn: () => void): string {
 }
 
 function newBaseProgress(items: string[], opts: { isTTY?: boolean; maxVisible?: number } = {}) {
-  const progress = new ConcurrentProgress({ ...opts, resultTitle: 'Results' });
+  const progress = new ConcurrentProgress({ ...opts, console: term, resultTitle: 'Results' });
   progress.setTotal(items.length);
   progress.addItems(items);
   return progress;
 }
 
 function newImportProgress(repos: string[], opts: { isTTY?: boolean; maxVisible?: number } = {}) {
-  const progress = new ImportProgress(opts);
+  const progress = new ImportProgress({ ...opts, console: term });
   progress.setTotal(repos.length);
   progress.addRepos(repos);
   return progress;
@@ -160,7 +162,7 @@ describe('ConcurrentProgress — TTY', () => {
 
 describe('ConcurrentProgress — options', () => {
   it('showResult: false still emits the item list in non-TTY', () => {
-    const progress = new ConcurrentProgress({ isTTY: false, showResult: false });
+    const progress = new ConcurrentProgress({ console: term, isTTY: false, showResult: false });
     progress.setTotal(1);
     progress.addItems(['item']);
     progress.start();
@@ -171,7 +173,7 @@ describe('ConcurrentProgress — options', () => {
   });
 
   it('showResult: false suppresses the Succeeded/Failed block in TTY', () => {
-    const progress = new ConcurrentProgress({ isTTY: true, showResult: false });
+    const progress = new ConcurrentProgress({ console: term, isTTY: true, showResult: false });
     progress.setTotal(1);
     progress.addItems(['item']);
     progress.start();
@@ -182,7 +184,11 @@ describe('ConcurrentProgress — options', () => {
   });
 
   it('resultTitle appears in TTY result header', () => {
-    const progress = new ConcurrentProgress({ isTTY: true, resultTitle: 'My Results' });
+    const progress = new ConcurrentProgress({
+      console: term,
+      isTTY: true,
+      resultTitle: 'My Results',
+    });
     progress.setTotal(1);
     progress.addItems(['item']);
     progress.start();
@@ -194,7 +200,7 @@ describe('ConcurrentProgress — options', () => {
 
 describe('ImportProgress — TTY', () => {
   it('setTotal fixes the bar denominator independently of addRepos, for a streaming job', () => {
-    const progress = new ImportProgress({ isTTY: true });
+    const progress = new ImportProgress({ console: term, isTTY: true });
     progress.setTotal(5);
 
     const start = captureStdout(() => progress.start());
@@ -217,7 +223,7 @@ describe('ImportProgress — TTY', () => {
   });
 
   it('a later page takes over rows already finished by an earlier page, instead of stalling in the queue', () => {
-    const progress = new ImportProgress({ isTTY: true, maxVisible: 2 });
+    const progress = new ImportProgress({ console: term, isTTY: true, maxVisible: 2 });
     progress.setTotal(4);
     progress.start();
 
@@ -241,7 +247,7 @@ describe('ImportProgress — TTY', () => {
   });
 
   it('recordSkipped advances the bar without adding a row', () => {
-    const progress = new ImportProgress({ isTTY: true });
+    const progress = new ImportProgress({ console: term, isTTY: true });
     progress.setTotal(3);
     progress.start();
